Treat FontForge error
There is a bug in pdf2htmlEX that makes its process consume all
available memory. This problem happens when FontForge, a dependency
of pdf2htmlEX, has an offset bigger than 65535 bytes on one lookup.
This error can be detected when pdf2htmlEX/FontForge prints on
stderr the following message:

Internal Error: Attempt to output 65744 into a 16-bit field. It will
be truncated and the file may not be useful.